Deleted unnecessary files, and removed multiple lines that were console logs and unused functions
