+++ README.md
@@ -3,8 +3,6 @@
 This a tool that does (...)
 
 ## Files
-### lsm_cost.js
-Most of the functions contained here have been copied to worker.js
 
 ### worker.js 
 For multithreading purposes a Web Worker has been implemented.
