+++ index.html
@@ -21,7 +21,6 @@
     <link href="css/tooltip.css" rel="stylesheet">
 
     <!-- Custom JS -->
-    <script src="js/lsm_bush_in_action.js"></script>
     <script src="js/autoRerun.js"></script>
     <script src="js/uploadDataFile.js"></script>
     <script src="js/uploadWorkloadFile.js"></script>
@@ -1208,11 +1207,7 @@
     <!-- Custom Theme JavaScript -->
     <script src="js/main.js"></script>
     <script src="js/show_more.js"></script>
-    <script src="js/lsm_cost.js"></script>
-	<script> setTimeout(init, 200);</script>
-</head>
-
-    <!-- Custom Theme JavaScript -->
+    <script> setTimeout(init, 200); </script>
     <script src="js/disappearMenu.js"></script>
 
     <!-- Start of StatCounter Code for Default Guide -->
